+++ src/index.js
@@ -1,6 +1,6 @@
 import rainbow from './rainbow/rainbowMethod.js';
-import { printf } from './formatters/LogFormatter.js';
-import settings from './formatters/LogFormatter.js';
+import printf from './formatters/LogFormatter.js';
+import settings from './formatters/settings.js';
 
 const charmlog = { rainbow, printf, settings };
